requirements.txt
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
click==8.1.7
dnspython==2.6.1
flasgger==0.9.7.1
Flask==3.0.2
Flask-Cors==4.0.0
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
mistune==3.0.2
packaging==24.0
pymongo==4.6.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
referencing==0.33.0
rpds-py==0.18.0
six==1.16.0
Werkzeug==3.0.1
